done all methods service without upload

+++ src/main/java/ru/codeportfolio/dto/ResourceResponseDto.java
@@ -5,7 +5,7 @@ import ru.codeportfolio.models.TypeFile;
 public record ResourceResponseDto(
         String path,
         String name,
-        Integer size,
+        Long size,
         TypeFile type
 ) {
 }
